requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.5 
certifi==2023.11.17
charset-normalizer==3.3.2
cytoolz==0.12.2
Django==4.2.8
django-adminlte3==0.1.6
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lru-dict==1.2.0
multidict==6.0.4
packaging==23.2
parsimonious==0.9.0
Pillow==10.1.0
protobuf==4.25.1
py-solc-x==2.0.2
pycryptodome==3.19.0
pyunormalize==15.1.0
pywin32==306
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rlp==4.0.0
rpds-py==0.13.2
sqlparse==0.4.4
toolz==0.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
web3==6.11.4
websockets==12.0
yarl==1.9.4
